services/backend-api/requirements.txt
fastapi[standard]>=0.115.0,<0.116.0
uvicorn>=0.27.0,<0.28.0
pydantic>=2.7.4,<2.10.0
pymongo>=4.3.3,<5.0.0
python-dotenv>=1.0.0,<2.0.0
passlib>=1.7.4,<2.0.0
python-jose>=3.3.0,<4.0.0
python-multipart>=0.0.5,<0.1.0
httpx>=0.28.1,<0.29.0
pytest>=7.3.1,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0
asyncio>=3.4.3,<4.0.0
beanie>=1.26.0,<2.0.0
moto>=4.1.0,<5.0.0
# Google API libraries for Google Docs validator
google-api-python-client>=2.86.0,<3.0.0
google-auth>=2.16.0,<3.0.0
google-auth-httplib2>=0.1.0,<1.0.0

# Data processing libraries
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
xlrd >= 2.0.1

# File handling libraries
aiofiles>=23.0.0,<24.0.0

# YAML parsing
PyYAML>=6.0.0,<7.0.0

# AWS libraries
boto3>=1.28.0,<2.0.0

# Message queue libraries
pika>=1.3.0,<2.0.0

# Server-Sent Events (SSE) support for streaming responses
sse-starlette>=1.6.0,<2.0.0

langchain-mcp-adapters==0.1.9

# Snowflake connector with compatible dependencies
snowflake-connector-python>=3.4.0,<4.0.0

# Cryptography for private key handling (required for Snowflake key pair authentication)
cryptography>=41.0.0,<43.0.0

# Google BigQuery libraries
google-cloud-bigquery>=3.0.0,<4.0.0

# Redash libraries
redash-toolbelt>=0.1.0

# DBT libraries
dbt-core>=1.10.9,<2.0.0

# Looker libraries
looker-sdk>=24.0.0,<25.0.0

# DataHub libraries
acryl-datahub>=0.13.0,<0.14.0

# Airflow libraries (for HTTP API client)
requests>=2.25.0,<3.0.0

# Redis Client
redis>=6.1.0,<7.0.0

# Azure libraries
azure-identity>=1.15.0,<2.0.0
azure-storage-blob>=12.19.0,<13.0.0
azure-mgmt-datafactory>=4.0.0,<5.0.0
